fix(sequence): expand sequence counter from u16 to u32 to support larger bit layouts
Change the internal sequence counter type from `u16` to `u32` to accommodate
bit layouts where `sequence_bits` exceeds 16 bits.

**Background**
- The sequence field can be configured up to 63 bits (though practically limited)
- Previously `u16` capped maximum sequence bits at 16 (max value 65535)
- This caused panics or truncation when using layouts with >16 sequence bits

**Changes**
- `SnowflakeIdGenerator.idx`: `u16` → `u32`
- `SnowflakeComponents.sequence`: `u16` → `u32`
- `BitLayout::max_sequence()` return type: `u16` → `u32`
- Update decomposition casting: `(id & sequence_mask) as u32`

**Impact**
- Supports sequence bits up to 32 (max value ~4.29 billion IDs per millisecond)
- Backward compatible for existing layouts (≤16 sequence bits still work)
- No breaking API changes for consumers using `max_sequence()` (only return type widens)

**Why**
- The `sequence_bits` field is `u8`, allowing configurations up to 255 bits
- While practical layouts use ≤32 sequence bits, `u16` artificially restricted valid configurations
- Aligns with the maximum sequence mask calculation which already used `u32` shifts

+++ src/generator/components.rs
@@ -10,5 +10,5 @@ pub struct SnowflakeComponents {
     /// Node identifier embedded in the ID.
     pub node_id: i32,
     /// Per-millisecond sequence counter embedded in the ID.
-    pub sequence: u16,
+    pub sequence: u32,
 }
